syscfg: update syscfg_set_u() to accept unsigned long instead of unsigned
A lot of RDKB code uses ULONG when it menas a 32bit unsigned value.
Update syscfg_set_u() so that it can accept the ULONG data type
without generating compiler warnings...

+++ source/include/syscfg/syscfg.h
@@ -143,13 +143,13 @@ int syscfg_getall(char *buf, int count, int *outsz);
 
 int syscfg_set_ns             (const char *ns, const char *name, const char *value);
 int syscfg_set_ns_commit      (const char *ns, const char *name, const char *value);
-int syscfg_set_ns_u           (const char *ns, const char *name, unsigned int value);
-int syscfg_set_ns_u_commit    (const char *ns, const char *name, unsigned int value);
+int syscfg_set_ns_u           (const char *ns, const char *name, unsigned long value);
+int syscfg_set_ns_u_commit    (const char *ns, const char *name, unsigned long value);
 
 int syscfg_set_nns            (const char *name, const char *value);
 int syscfg_set_nns_commit     (const char *name, const char *value);
-int syscfg_set_nns_u          (const char *name, unsigned int value);
-int syscfg_set_nns_u_commit   (const char *name, unsigned int value);
+int syscfg_set_nns_u          (const char *name, unsigned long value);
+int syscfg_set_nns_u_commit   (const char *name, unsigned long value);
 
 static inline int syscfg_set (const char *ns, const char *name, const char *value)
 {
@@ -167,7 +167,7 @@ static inline int syscfg_set_commit (const char *ns, const char *name, const cha
         return syscfg_set_nns_commit (name, value);
 }
 
-static inline int syscfg_set_u (const char *ns, const char *name, unsigned int value)
+static inline int syscfg_set_u (const char *ns, const char *name, unsigned long value)
 {
     if (ns)
         return syscfg_set_ns_u (ns, name, value);
@@ -175,7 +175,7 @@ static inline int syscfg_set_u (const char *ns, const char *name, unsigned int v
         return syscfg_set_nns_u (name, value);
 }
 
-static inline int syscfg_set_u_commit (const char *ns, const char *name, unsigned int value)
+static inline int syscfg_set_u_commit (const char *ns, const char *name, unsigned long value)
 {
     if (ns)
         return syscfg_set_ns_u_commit (ns, name, value);
